+++ lib/app_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:projeto_criptos/shared/theme/theme_data_criptos.dart';
+import 'shared/theme/theme_data_criptos.dart';
 import 'route_controller.dart';
 
 class AppWidget extends StatelessWidget {
