+++ lib/conversion/logicholder/methods/show_modal_bottom_sheet_cryptos.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:projeto_criptos/conversion/model/conversion_crypto_view_data.dart';
+import '../../../shared/common_model/crypto.dart';
 
 import '../../../shared/templates/app_assets.dart';
 
 Future<void> showModalBottomSheetCryptos(
-    BuildContext context, List<ConversionCryptoViewData> data, Widget listView) {
+    BuildContext context, List<CryptoEntity> data, Widget listView) {
   return showModalBottomSheet<void>(
     context: context,
     shape: const RoundedRectangleBorder(
