+++ lib/conversion/view/conversion_screen.dart
@@ -1,11 +1,10 @@
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/conversion/controller/controller_arguments.dart';
-import 'package:projeto_criptos/conversion/controller/conversion_cryptos_provider.dart';
-import 'package:projeto_criptos/conversion/model/conversion_crypto_view_data.dart';
-import 'package:projeto_criptos/conversion/widgets/total_convert_value_container.dart';
-import 'package:projeto_criptos/shared/common_model/crypto.dart';
+import '../../portfolio/model/crypto_view_data.dart';
+import '../controller/controller_arguments.dart';
+import '../widgets/total_convert_value_container.dart';
+import '../../shared/common_model/crypto.dart';
 import '../../shared/user/user_coin_ammount_provider.dart';
 import '../controller/converted_crypto_provider.dart';
 import '../controller/validate_provider.dart';
@@ -15,8 +14,6 @@ import '../widgets/coin_button.dart';
 import '../widgets/coin_text_field.dart';
 import '../widgets/swap_icon_button.dart';
 import '../../portfolio/controller/cryptos_provider.dart';
-import '../../shared/templates/error_body.dart';
-import '../../shared/templates/loading_body.dart';
 
 import '../../shared/utils/decimal_parse.dart';
 import '../logicholder/methods/show_modal_bottom_sheet_cryptos.dart';
@@ -71,173 +68,158 @@ class _$ConversionScreenState extends ConsumerState<ConversionScreen> {
 
   @override
   Widget build(BuildContext context) {
-    final cryptos = ref.watch(conversionCryptosProvider);
+    final cryptos = ref.watch(cryptosProvider);
     final controller = ref.watch(controllerArgumentsProvider.state).state;
     controller.convertCoin = asset;
     controller.recieveCoin = cryptoConverted;
-    return cryptos.when(
-      data: (data) {
-        if (cryptoConverted.id == 'id') {
-          cryptoConverted = data[0].id == asset.id ? data[1] : data[0];
-        }
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
+    if (cryptoConverted.id == 'id') {
+      cryptoConverted = cryptos.asData!.value[0].id == asset.id
+          ? cryptos.asData!.value[1]
+          : cryptos.asData!.value[0];
+    }
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        AvailableBalanceContainer(
+          asset: asset,
+          coinAmmount: coinAmmount,
+        ),
+        const InformativeText(),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
           children: [
-            AvailableBalanceContainer(
-              asset: asset,
-              coinAmmount: coinAmmount,
-            ),
-            const InformativeText(),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: [
-                CoinButton(
-                  data: data,
-                  onPressed: () {
-                    showModalBottomSheetCryptos(
-                      context,
-                      data,
-                      ListView(
-                        children: data.map((crypto) {
-                          int index = data.indexOf(crypto);
-                          return Column(
-                            children: [
-                              const Divider(thickness: 1),
-                              ListTile(
-                                onTap: () {
-                                  setState(() {
-                                    if (cryptoConverted == crypto) {
-                                      cryptoConverted = asset;
-                                      controller.recieveCoin = cryptoConverted;
-                                    }
-                                    asset = crypto;
-                                    coinAmmount = dp(ref
-                                        .read(userCoinAmmountProvider)[index]
-                                        .toString());
-                                    controller.convertCoin = asset;
-                                  });
-                                  convertController.clear();
-                                  buttonValidation();
-                                  convertedValue('0');
-                                  Navigator.of(context).pop();
-                                },
-                                title: Text(
-                                  crypto.symbol.toUpperCase(),
-                                  style: const TextStyle(
-                                    fontWeight: FontWeight.bold,
-                                  ),
-                                ),
-                                subtitle: Text(crypto.name),
-                                trailing:
-                                    const Icon(Icons.keyboard_arrow_right),
+            CoinButton(
+              data: cryptos.asData!.value,
+              onPressed: () {
+                showModalBottomSheetCryptos(
+                  context,
+                  cryptos.asData!.value,
+                  ListView(
+                    children: cryptos.asData!.value.map((crypto) {
+                      int index = cryptos.asData!.value.indexOf(crypto);
+                      return Column(
+                        children: [
+                          const Divider(thickness: 1),
+                          ListTile(
+                            onTap: () {
+                              setState(() {
+                                if (cryptoConverted == crypto) {
+                                  cryptoConverted = asset;
+                                  controller.recieveCoin = cryptoConverted;
+                                }
+                                asset = crypto;
+                                coinAmmount = dp(ref
+                                    .read(userCoinAmmountProvider)[index]
+                                    .toString());
+                                controller.convertCoin = asset;
+                              });
+                              convertController.clear();
+                              buttonValidation();
+                              convertedValue('0');
+                              Navigator.of(context).pop();
+                            },
+                            title: Text(
+                              crypto.symbol.toUpperCase(),
+                              style: const TextStyle(
+                                fontWeight: FontWeight.bold,
                               ),
-                            ],
-                          );
-                        }).toList(),
-                      ),
-                    );
-                  },
-                  asset: asset,
-                ),
-                SwapIconButton(
-                  onPressed: () {
-                    CryptoEntity temp = asset;
-                    setState(() {
-                      asset = cryptoConverted;
-                      controller.convertCoin = asset;
-                      cryptoConverted = temp;
-                      controller.recieveCoin = cryptoConverted;
-                      coinAmmount = dp(ref
-                          .read(userCoinAmmountProvider)[
-                              data.indexOf(asset as ConversionCryptoViewData)]
-                          .toString());
-                    });
-                    convertController.clear();
-                    buttonValidation();
-                    convertedValue('0.0');
-                  },
-                ),
-                CoinButton(
-                  data: data,
-                  onPressed: () {
-                    showModalBottomSheetCryptos(
-                      context,
-                      data,
-                      ListView(
-                        children: data.map((crypto) {
-                          return Column(
-                            children: [
-                              const Divider(thickness: 1),
-                              ListTile(
-                                onTap: () {
-                                  CryptoEntity temp = cryptoConverted;
-                                  setState(() {
-                                    cryptoConverted = crypto;
-                                    controller.recieveCoin = cryptoConverted;
-                                    if (cryptoConverted.id == asset.id) {
-                                      asset = temp;
-                                      controller.convertCoin = asset;
-                                    }
-                                  });
-                                  buttonValidation();
-                                  convertedValue(convertController.text);
-                                  Navigator.of(context).pop();
-                                },
-                                title: Text(crypto.symbol.toUpperCase()),
-                                subtitle: Text(crypto.name),
-                                trailing:
-                                    const Icon(Icons.keyboard_arrow_right),
-                              ),
-                            ],
-                          );
-                        }).toList(),
-                      ),
-                    );
-                  },
-                  asset: cryptoConverted,
-                ),
-              ],
-            ),
-            CoinTextField(
-              formKey: formKey,
-              controller: convertController,
+                            ),
+                            subtitle: Text(crypto.name),
+                            trailing: const Icon(Icons.keyboard_arrow_right),
+                          ),
+                        ],
+                      );
+                    }).toList(),
+                  ),
+                );
+              },
               asset: asset,
-              onChanged: (value) {
+            ),
+            SwapIconButton(
+              onPressed: () {
+                CryptoEntity temp = asset;
+                setState(() {
+                  asset = cryptoConverted;
+                  controller.convertCoin = asset;
+                  cryptoConverted = temp;
+                  controller.recieveCoin = cryptoConverted;
+                  coinAmmount = dp(ref
+                      .read(userCoinAmmountProvider)[cryptos.asData!.value
+                          .indexOf(asset as CryptoViewData)]
+                      .toString());
+                });
+                convertController.clear();
                 buttonValidation();
-                convertedValue(value);
-                controller.convert = dp(value);
-                controller.recieve = convertedCryptoHelper;
-              },
-              validator: (value) {
-                if (value == '' || value == null || double.tryParse(value) == 0) {
-                  return 'Valor deve ser maior que zero';
-                } else if (ConversionMethods.validCoinValue(value)) {
-                  return 'O valor inicial não pode ser um caractere especial';
-                } else if (dp(ConversionMethods.coinRegExp(value)) >
-                    coinAmmount) {
-                  return 'Saldo Insuficiente';
-                }
-                return null;
+                convertedValue('0.0');
               },
             ),
-            HelperCurrencyText(convertHelper: convertHelper),
-            TotalConvertValueContainer(
-              convertedCryptoHelper: convertedCryptoHelper,
-              cryptoConverted: cryptoConverted,
+            CoinButton(
+              data: cryptos.asData!.value,
+              onPressed: () {
+                showModalBottomSheetCryptos(
+                  context,
+                  cryptos.asData!.value,
+                  ListView(
+                    children: cryptos.asData!.value.map((crypto) {
+                      return Column(
+                        children: [
+                          const Divider(thickness: 1),
+                          ListTile(
+                            onTap: () {
+                              CryptoEntity temp = cryptoConverted;
+                              setState(() {
+                                cryptoConverted = crypto;
+                                controller.recieveCoin = cryptoConverted;
+                                if (cryptoConverted.id == asset.id) {
+                                  asset = temp;
+                                  controller.convertCoin = asset;
+                                }
+                              });
+                              buttonValidation();
+                              convertedValue(convertController.text);
+                              Navigator.of(context).pop();
+                            },
+                            title: Text(crypto.symbol.toUpperCase()),
+                            subtitle: Text(crypto.name),
+                            trailing: const Icon(Icons.keyboard_arrow_right),
+                          ),
+                        ],
+                      );
+                    }).toList(),
+                  ),
+                );
+              },
+              asset: cryptoConverted,
             ),
           ],
-        );
-      },
-      error: (e, s) {
-        return ErrorBody(
-          onRetry: () {
-            ref.refresh(cryptosProvider);
+        ),
+        CoinTextField(
+          formKey: formKey,
+          controller: convertController,
+          asset: asset,
+          onChanged: (value) {
+            buttonValidation();
+            convertedValue(value);
+            controller.convert = dp(value);
+            controller.recieve = convertedCryptoHelper;
           },
-        );
-      },
-      loading: () {
-        return const LoadingBody();
-      },
+          validator: (value) {
+            if (value == '' || value == null || double.tryParse(value) == 0) {
+              return 'Valor deve ser maior que zero';
+            } else if (ConversionMethods.validCoinValue(value)) {
+              return 'O valor inicial não pode ser um caractere especial';
+            } else if (dp(ConversionMethods.coinRegExp(value)) > coinAmmount) {
+              return 'Saldo Insuficiente';
+            }
+            return null;
+          },
+        ),
+        HelperCurrencyText(convertHelper: convertHelper),
+        TotalConvertValueContainer(
+          convertedCryptoHelper: convertedCryptoHelper,
+          cryptoConverted: cryptoConverted,
+        ),
+      ],
     );
   }
 }
