+++ lib/conversion/widgets/floating_cripto_button.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/shared/utils/arguments/to_revision_arguments.dart';
-import '../controller/controller_arguments.dart';
-import '../controller/validate_provider.dart';
 
 import '../../shared/templates/app_assets.dart';
+import '../../shared/utils/arguments/to_revision_arguments.dart';
+import '../controller/controller_arguments.dart';
+import '../controller/validate_provider.dart';
 
 class FloatingCriptoButton extends HookConsumerWidget {
   const FloatingCriptoButton({
