+++ lib/details/widgets/info_column.dart
@@ -2,10 +2,10 @@ import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 
 import '../../shared/common_model/crypto.dart';
+import '../../shared/templates/default_info_row.dart';
 import '../../shared/utils/currency_formater.dart';
 import '../../shared/utils/decimal_to_double.dart';
 import '../model/prices_view_data.dart';
-import '../../shared/templates/default_info_row.dart';
 
 class InfoColumn extends StatelessWidget {
   const InfoColumn({
