+++ lib/details/widgets/top_page_container.dart
@@ -1,11 +1,11 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../controller/get_price_provider.dart';
-import '../../shared/utils/decimal_to_double.dart';
-import '../../shared/common_model/crypto.dart';
 
+import '../../shared/common_model/crypto.dart';
 import '../../shared/utils/currency_formater.dart';
+import '../../shared/utils/decimal_to_double.dart';
+import '../controller/get_price_provider.dart';
 
 class TopPageContainer extends HookConsumerWidget {
   const TopPageContainer({
