+++ lib/movements/view/moves_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/shared/user/movements_provider.dart';
-import 'package:projeto_criptos/shared/templates/app_assets.dart';
+import '../../shared/user/movements_provider.dart';
+import '../../shared/templates/app_assets.dart';
 
 import '../widgets/movements_list_tile.dart';
