+++ lib/portfolio/controller/cryptos_provider.dart
@@ -1,6 +1,6 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:projeto_criptos/portfolio/usecase/cryptos_usecase.dart';
-import 'package:projeto_criptos/shared/repository/cryptos_repository_provider.dart';
+import '../usecase/cryptos_usecase.dart';
+import '../../shared/repository/cryptos_repository_provider.dart';
 
 import '../model/crypto_view_data.dart';
