+++ lib/shared/utils/arguments/to_revision_arguments.dart
@@ -1,6 +1,6 @@
 import 'package:decimal/decimal.dart';
 
-import 'package:projeto_criptos/shared/common_model/crypto.dart';
+import '../../common_model/crypto.dart';
 
 class ToRevisionArguments {
   Decimal convert;
